Turned some of the process functions into separate files.
